app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
Pillow==10.3.0
opencv-python-headless==4.10.0.82
ultralytics==8.2.23
shapely==2.0.4
pyshp==2.3.1
jinja2==3.1.4
supabase==2.5.3
python-dotenv==1.0.1